Support Redux DevTools Extension in dashboard v2spa.
It is currently not possible to use the Redux DevTools Extension with v2spa.
Redux DevTools Extension allows us to view the Redux state historically before
and after each action dispatched. This is incredibly helpful for debugging.
This CL allows us to use Redux DevTools Extension in development and
production.

Performance:
The Redux DevTools Extension is supported through Redux middleware, which is
only injected when the user has Redux DevTools installed.  This will not cause
a performance hit for users who do not have the extension installed. For users
who do have the extension installed in production, the extension is turned off
by default. To turn the extension on, users can click the "Start recording"
button.

Demos:
 - https://redux-devtools-extension-dot-chromeperf.appspot.com/
 - https://redux-devtools-extension-prod-dot-chromeperf.appspot.com/

Change-Id: I07eef459052fa4b66248bf284d60a2bb0173d006
Bug: chromium:855252
Reviewed-on: https://chromium-review.googlesource.com/1108570
Commit-Queue: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>