Cache report data in the Redux store.
Performance reports were not stored in local memory, causing the
application to refetch a report even if it was already loaded in the
same session. These changes add an intermediate step in which network
data is cached by Redux.  This makes revistiting reports seemingly
instantaneous.

The next step would be adding IndexedDB for a persistant cache between
sessions and even tabs. Track the progress of this in CL 1176384:
https://chromium-review.googlesource.com/c/catapult/+/1176384

Demo: https://reports-redux-dot-chromeperf.appspot.com

Bug: chromium:855252
Change-Id: Ia49fad387bc9b3737fec3ff1e6a6284c264d5c7a
Reviewed-on: https://chromium-review.googlesource.com/1176726
Commit-Queue: Sam Balana <sbalana@google.com>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>